+++ plugins/11_TheWarWithin/zones/hallowfall.lua
@@ -1292,3 +1292,18 @@ map.nodes[61501747] = BeaconOfHope({
         })
     }
 })
+
+-------------------------------------------------------------------------------
+------------------------ ACHIEVEMENT: FLAMEGARD'S HOPE ------------------------
+-------------------------------------------------------------------------------
+
+map.nodes[43205177] = Collectible({
+    icon = 463526,
+    quest = 79081,
+    label = '{achievement:20594}',
+    note = L['flamegards_hope_note'],
+    rewards = {
+        Achievement(
+            {id = 20594, criteria = {id = 1, qty = true}})
+    }
+})